Fixed Rectangle
Fixed Rectangle. There was an incorrect assignment of a point to the
point data member of the Rectangle class. I fixed this, and then
refactored the code to use more descriptive variable and data member
names to avoid such small typos and confusion. After these changes code
passed all tests.